test(frontend): move useSettings llm base url test
